data_producer_app/requirements.txt
kafka-python
browser_cookie3
requests